python-tools/requirements.txt
numpy
openai
openpyxl
pandas
google-generativeai
